+++ qadence-commons/README.md
@@ -1,20 +1,20 @@
-# Qadence-Common
+# Qadence-Commons
 
-**Qadence-Common** is a Python package that provides shared libraries for qadence-hub. **Common** documentations is [here](https://pasqal-io.github.io/qadence-hub/qadence-common/latest/)
+**Qadence-Commons** is a Python package that provides shared libraries for qadence-hub. **Commons** documentations is [here](https://pasqal-io.github.io/qadence-hub/qadence-commons/latest/)
 
 [![Linting](https://github.com/pasqal-io/qadence-hub/actions/workflows/lint.yml/badge.svg)](https://github.com/pasqal-io/qadence-hub/actions/workflows/lint.yml)
 [![Tests](https://github.com/pasqal-io/qadence-hub/actions/workflows/test_fast.yml/badge.svg)](https://github.com/pasqal-io/qadence-hub/actions/workflows/test.yml)
 [![Documentation](https://github.com/pasqal-io/qadence-hub/actions/workflows/build_docs.yml/badge.svg)](https://pasqal-io.github.io/qadence-hub/latest)
-[![Pypi](https://badge.fury.io/py/qadence-common.svg)](https://pypi.org/project/qadence-common/)
+[![Pypi](https://badge.fury.io/py/qadence-commons.svg)](https://pypi.org/project/qadence-commons/)
 [![License](https://img.shields.io/badge/License-Apache_2.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
 
 
 ## Installation guide
 
-[qadence-common](https://pypi.org/project/qadence-common/) and can be installed using `pip` as follows:
+[qadence-commons](https://pypi.org/project/qadence-commons/) and can be installed using `pip` as follows:
 
 ```bash
-pip install qadence-common
+pip install qadence-commons
 ```
 
 ## Contributing
@@ -22,7 +22,7 @@ pip install qadence-common
 Before making a contribution, please review our [code of conduct](docs/CODE_OF_CONDUCT.md).
 
 - **Submitting Issues:** To submit bug reports or feature requests, please use our [issue tracker](https://github.com/pasqal-io/qadence-hub/issues).
-- **Developing in qadence:** To learn more about how to develop within `qadence-common`, please refer to [contributing guidelines](docs/CONTRIBUTING.md).
+- **Developing in qadence:** To learn more about how to develop within `qadence-commons`, please refer to [contributing guidelines](docs/CONTRIBUTING.md).
 
 ### Setting up qadence in development mode
 
@@ -49,4 +49,4 @@ python -m pip install -e .
 
 
 ## License
-Qadence-Common is a free and open source software package, released under the Apache License, Version 2.0.
+Qadence-Commons is a free and open source software package, released under the Apache License, Version 2.0.
