+++ qadence-commons/qadence_commons/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
-from qadence_common.protocol import Protocol, StrEnum
+from qadence_commons.protocol import Protocol, StrEnum
 
 __all__ = ["Protocol", "StrEnum"]
